Fix bad delta_height on subsequent loop iterations
The delta_height was always the same, because it did not consider the last known timestamp of the last block fetched in the previous iteration of the loop.

Note that the v1 part (blocktime 60s) does not have this issue, because it always calculates delta_height from the last known timestamp.

When falling in the second case (timestamp > xmr_blocktime_120), the first iteration must calculate the delta from the fork timestamp, while subsequent iterations must calculate.